Remove disposition column from call table migration

+++ database/migrations/2024_02_19_115726_remove_disposition_column_from_call_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('calls', function (Blueprint $table) {
-            $table->dropColumn('disposition');
+            if (Schema::hasColumn('calls', 'disposition')) { // Check if the column exists
+                $table->dropColumn('disposition');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('calls', function (Blueprint $table) {
-            $table->string('disposition')->nullable()->after('call_type_id');
+            if (!Schema::hasColumn('calls', 'disposition')) { // Check if the column doesn't exist
+                $table->string('disposition')->nullable()->after('call_type_id');
+            }
         });
     }
 };
